fix(export): tighten architecture package contract
